requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
apscheduler==3.10.4
gspread==6.1.4
google-auth==2.37.0
httpx==0.28.1
pydantic-settings==2.7.0
